refactor InteractiveModeInput to only show search input
Previously, InteractiveModeInput also was responsible (in some cases) for showing the global navbar's logo and right-hand nav links and user avatar. That was because we wanted the interactive mode's selected filters row and add-filter row to extend full-width in the global navbar, which meant that it needed to render the entire global navbar. That is not necessary, and eliminating that greatly simplifies the implementation.